refactor: 이벤트 리스너를 각 컴포넌트로 분리
