+++ public/js/views/blocks/dow-serialization.js
@@ -1,6 +1,6 @@
 import { dowNav } from "../components/dow-nav.js";
 import { dowSerializationNav } from "../components/dow-serialization-nav.js";
-import { dowTopWebtoons } from "../../controllers/fixing-webtoons.js";
+import { dowTopWebtoons } from "../../../../fixing-webtoons.js";
 import { webtoonCard } from "../components/webtoon-card.js";
 
 const position = "beforeend";
@@ -77,5 +77,11 @@ const dowSerialization = () => {
   containerTop.init();
   return dowNav() + dowSerializationNav() + dowSerializationContents();
 };
+
+const setAdBannerEvent = () => {
+  window.addEventListener("scroll", scrollDowSerialization);
+};
+
+setAdBannerEvent();
 
 export { scrollDowSerialization, dowSerialization };
